+++ R/shorthand.R
@@ -5,8 +5,9 @@
 #' @param x Value to be converted
 #' @param ... other args for as. conversion
 #'
-#' # Shorthand aliases for common conversions
-#' # Nothing magical here, but it can make your code more readable
+#' @note
+#' These are shorthand aliases for common conversions There is nothing magical
+#' here, but it can make your code more readable
 #'
 #' @examples
 #' chr(42)   # "42" = as.character